Refs #20817: Remove rtps_fwd.h
